+++ src/com/maddyhome/idea/vim/RegisterActions.java
@@ -17,10 +17,13 @@
  */
 package com.maddyhome.idea.vim;
 
+import com.intellij.openapi.extensions.ExtensionPointListener;
 import com.intellij.openapi.extensions.ExtensionPointName;
+import com.intellij.openapi.extensions.PluginDescriptor;
 import com.maddyhome.idea.vim.group.KeyGroup;
 import com.maddyhome.idea.vim.handler.ActionBeanClass;
 import com.maddyhome.idea.vim.handler.EditorActionHandlerBase;
+import com.maddyhome.idea.vim.key.MappingOwner;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -31,17 +34,41 @@ public class RegisterActions {
 
   public static final ExtensionPointName<ActionBeanClass> VIM_ACTIONS_EP =
     ExtensionPointName.create("IdeaVIM.vimAction");
-  private static boolean actionsRegistered = false;
+  private static boolean initialRegistration = false;
+
+  static {
+    // IdeaVim doesn't support contribution to VIM_ACTIONS_EP extension point, so technically we can skip this update,
+    //   but let's support dynamic plugins in a more classic way and reload actions on every EP change.
+    // TODO: [VERSION UPDATE] since 191 use
+    //  ExtensionPoint.addExtensionPointListener(ExtensionPointListener<T>, boolean, Disposable)
+    // TODO: [VERSION UPDATE] since 201 use
+    //  ExtensionPoint.addExtensionPointListener(ExtensionPointChangeListener, boolean, Disposable)
+    VIM_ACTIONS_EP.getPoint(null).addExtensionPointListener(new ExtensionPointListener<ActionBeanClass>() {
+      @Override
+      public void extensionAdded(@NotNull ActionBeanClass extension, @NotNull PluginDescriptor pluginDescriptor) {
+        // Suppress listener before the `VimPlugin.turnOn()` function execution. This logic should be rewritten after
+        //   version update (or earlier).
+        if (!initialRegistration) return;
+        unregisterActions();
+        registerActions();
+      }
+
+      @Override
+      public void extensionRemoved(@NotNull ActionBeanClass extension, @NotNull PluginDescriptor pluginDescriptor) {
+        if (!initialRegistration) return;
+        unregisterActions();
+        registerActions();
+      }
+    });
+  }
 
   /**
    * Register all the key/action mappings for the plugin.
    */
-  static void registerActions() {
-    if (actionsRegistered) return;
-    actionsRegistered = true;
-
+  public static void registerActions() {
     registerVimCommandActions();
     registerEmptyShortcuts();
+    initialRegistration = true;
   }
 
   @Nullable
@@ -57,6 +84,10 @@ public class RegisterActions {
     return action;
   }
 
+  public static void unregisterActions() {
+    VimPlugin.getKey().unregisterCommandActions();
+  }
+
   private static void registerVimCommandActions() {
     KeyGroup parser = VimPlugin.getKey();
     VIM_ACTIONS_EP.extensions().forEach(parser::registerCommandAction);
@@ -67,6 +98,6 @@ public class RegisterActions {
 
     // The {char1} <BS> {char2} shortcut is handled directly by KeyHandler#handleKey, so doesn't have an action. But we
     // still need to register the shortcut, to make sure the editor doesn't swallow it.
-    parser.registerShortcutWithoutAction(KeyStroke.getKeyStroke(KeyEvent.VK_BACK_SPACE, 0));
+    parser.registerShortcutWithoutAction(KeyStroke.getKeyStroke(KeyEvent.VK_BACK_SPACE, 0), MappingOwner.IdeaVim.INSTANCE);
   }
 }