+++ src/com/maddyhome/idea/vim/StatusBar.kt
@@ -50,6 +50,7 @@ import com.intellij.ui.awt.RelativePoint
 import com.intellij.util.Consumer
 import com.intellij.util.text.VersionComparatorUtil
 import com.maddyhome.idea.vim.group.NotificationService
+import com.maddyhome.idea.vim.option.IdeaStatusIcon
 import com.maddyhome.idea.vim.option.OptionsManager
 import com.maddyhome.idea.vim.ui.VimEmulationConfigurable
 import icons.VimIcons
@@ -59,7 +60,11 @@ import javax.swing.Icon
 import javax.swing.SwingConstants
 
 private class StatusBarIconProvider : StatusBarWidgetProvider {
-  override fun getWidget(project: Project): VimStatusBar? = if (OptionsManager.ideastatusbar.isSet) VimStatusBar else null
+  override fun getWidget(project: Project): VimStatusBar? {
+    if (!OptionsManager.ideastatusbar.isSet) return null
+    if (OptionsManager.ideastatusicon.value == IdeaStatusIcon.disabled) return null
+    return VimStatusBar
+  }
 }
 
 object VimStatusBar : StatusBarWidget, StatusBarWidget.IconPresentation {
@@ -76,7 +81,10 @@ object VimStatusBar : StatusBarWidget, StatusBarWidget.IconPresentation {
 
   override fun getTooltipText() = "IdeaVim"
 
-  override fun getIcon(): Icon = if (VimPlugin.isEnabled()) VimIcons.IDEAVIM else VimIcons.IDEAVIM_DISABLED
+  override fun getIcon(): Icon {
+    if (OptionsManager.ideastatusicon.value == IdeaStatusIcon.gray) return VimIcons.IDEAVIM_DISABLED
+    return if (VimPlugin.isEnabled()) VimIcons.IDEAVIM else VimIcons.IDEAVIM_DISABLED
+  }
 
   override fun getClickConsumer() = Consumer<MouseEvent> { event ->
     val component = event.component