+++ src/com/maddyhome/idea/vim/VimPlugin.java
@@ -331,6 +331,8 @@ public class VimPlugin implements BaseComponent, PersistentStateComponent<Elemen
   }
 
   public static void setEnabled(final boolean enabled) {
+    if (isEnabled() == enabled) return;
+
     if (!enabled) {
       getInstance().turnOffPlugin();
     }
@@ -425,6 +427,12 @@ public class VimPlugin implements BaseComponent, PersistentStateComponent<Elemen
 
   private void turnOffPlugin() {
     KeyHandler.getInstance().fullReset(null);
+
+    // Unregister vim actions in command mode
+    RegisterActions.unregisterActions();
+
+    // Unregister ex handlers
+    CommandParser.getInstance().unregisterHandlers();
 
     getEditor().turnOff();
     getSearch().turnOff();