+++ src/com/maddyhome/idea/vim/extension/VimExtensionRegistrar.java
@@ -21,6 +21,7 @@ package com.maddyhome.idea.vim.extension;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.extensions.ExtensionPointListener;
 import com.intellij.openapi.extensions.PluginDescriptor;
+import com.maddyhome.idea.vim.key.MappingOwner;
 import com.maddyhome.idea.vim.option.OptionsManager;
 import com.maddyhome.idea.vim.option.ToggleOption;
 import org.jetbrains.annotations.NotNull;
@@ -34,8 +35,10 @@ import java.util.Set;
  */
 public class VimExtensionRegistrar {
 
-  private static Set<String> registeredExtensions = new HashSet<>();
+  private static final Set<String> registeredExtensions = new HashSet<>();
   private static boolean extensionRegistered = false;
+
+  private static final Logger logger = Logger.getInstance(VimExtensionRegistrar.class);
 
   public static void registerExtensions() {
     if (extensionRegistered) return;
@@ -43,10 +46,16 @@ public class VimExtensionRegistrar {
 
     // TODO: [VERSION UPDATE] since 191 use
     //  ExtensionPoint.addExtensionPointListener(ExtensionPointListener<T>, boolean, Disposable)
+    //noinspection deprecation
     VimExtension.EP_NAME.getPoint(null).addExtensionPointListener(new ExtensionPointListener<VimExtension>() {
       @Override
       public void extensionAdded(@NotNull VimExtension extension, @NotNull PluginDescriptor pluginDescriptor) {
         registerExtension(extension);
+      }
+
+      @Override
+      public void extensionRemoved(@NotNull VimExtension extension, @NotNull PluginDescriptor pluginDescriptor) {
+        unregisterExtension(extension);
       }
     });
   }
@@ -74,5 +83,15 @@ public class VimExtensionRegistrar {
     OptionsManager.INSTANCE.addOption(option);
   }
 
-  private static Logger logger = Logger.getInstance(VimExtensionRegistrar.class);
+  synchronized private static void unregisterExtension(@NotNull VimExtension extension) {
+    String name = extension.getName();
+
+    if (!registeredExtensions.contains(name)) return;
+
+    registeredExtensions.remove(name);
+    extension.dispose();
+    OptionsManager.INSTANCE.removeOption(name);
+    MappingOwner.Plugin.Companion.remove(name);
+    logger.info("IdeaVim extension '" + name + "' disposed");
+  }
 }