+++ src/com/maddyhome/idea/vim/key/KeyMapping.java
@@ -20,13 +20,14 @@ package com.maddyhome.idea.vim.key;
 
 import com.google.common.collect.HashMultiset;
 import com.google.common.collect.Multiset;
-import com.maddyhome.idea.vim.command.MappingMode;
 import com.maddyhome.idea.vim.extension.VimExtensionHandler;
+import kotlin.Pair;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
 import java.util.*;
+import java.util.stream.Collectors;
 
 /**
  * Container for key mappings for some mode
@@ -35,7 +36,9 @@ import java.util.*;
  * @author vlan
  */
 public class KeyMapping implements Iterable<List<KeyStroke>> {
-  /** Contains all key mapping for some mode. */
+  /**
+   * Contains all key mapping for some mode.
+   */
   @NotNull private final Map<List<KeyStroke>, MappingInfo> myKeys = new HashMap<>();
   /**
    * Set the contains all possible prefixes for mappings.
@@ -58,13 +61,23 @@ public class KeyMapping implements Iterable<List<KeyStroke>> {
     return myKeys.get(keys);
   }
 
-  public void put(@NotNull Set<MappingMode> mappingModes,
-                  @NotNull List<KeyStroke> fromKeys,
-                  @Nullable List<KeyStroke> toKeys,
-                  @Nullable VimExtensionHandler extensionHandler,
+  public void put(@NotNull List<KeyStroke> fromKeys,
+                  @NotNull MappingOwner owner,
+                  @NotNull VimExtensionHandler extensionHandler,
                   boolean recursive) {
-    myKeys.put(new ArrayList<>(fromKeys),
-               new MappingInfo(mappingModes, fromKeys, toKeys, extensionHandler, recursive));
+    myKeys.put(new ArrayList<>(fromKeys), new ToHandlerMappingInfo(extensionHandler, fromKeys, recursive, owner));
+    fillPrefixes(fromKeys);
+  }
+
+  public void put(@NotNull List<KeyStroke> fromKeys,
+                  @NotNull List<KeyStroke> toKeys,
+                  @NotNull MappingOwner owner,
+                  boolean recursive) {
+    myKeys.put(new ArrayList<>(fromKeys), new ToKeysMappingInfo(toKeys, fromKeys, recursive, owner));
+    fillPrefixes(fromKeys);
+  }
+
+  private void fillPrefixes(@NotNull List<KeyStroke> fromKeys) {
     List<KeyStroke> prefix = new ArrayList<>();
     final int prefixLength = fromKeys.size() - 1;
     for (int i = 0; i < prefixLength; i++) {
@@ -73,14 +86,24 @@ public class KeyMapping implements Iterable<List<KeyStroke>> {
     }
   }
 
-  public void delete(@NotNull List<KeyStroke> keys) {
-    myKeys.remove(keys);
-    List<KeyStroke> prefix = new ArrayList<>();
-    final int prefixLength = keys.size() - 1;
-    for (int i = 0; i < prefixLength; i++) {
-      prefix.add(keys.get(i));
-      myPrefixes.remove(prefix);
-    }
+  public void delete(@NotNull MappingOwner owner) {
+    List<Map.Entry<List<KeyStroke>, MappingInfo>> toRemove =
+      myKeys.entrySet().stream().filter(o -> o.getValue().getOwner().equals(owner)).collect(Collectors.toList());
+
+    toRemove.forEach(o -> myKeys.remove(o.getKey(), o.getValue()));
+    toRemove.stream().map(Map.Entry::getKey).forEach(keys -> {
+      List<KeyStroke> prefix = new ArrayList<>();
+      final int prefixLength = keys.size() - 1;
+      for (int i = 0; i < prefixLength; i++) {
+        prefix.add(keys.get(i));
+        myPrefixes.remove(prefix);
+      }
+    });
+  }
+
+  public List<Pair<List<KeyStroke>, MappingInfo>> getByOwner(@NotNull MappingOwner owner) {
+    return myKeys.entrySet().stream().filter(o -> o.getValue().getOwner().equals(owner))
+      .map(o -> new Pair<>(o.getKey(), o.getValue())).collect(Collectors.toList());
   }
 
   public boolean isPrefix(@NotNull Iterable<KeyStroke> keys) {