+++ src/com/maddyhome/idea/vim/option/OptionsManager.kt
@@ -36,6 +36,7 @@ import com.maddyhome.idea.vim.helper.isBlockCaret
 import com.maddyhome.idea.vim.helper.mode
 import com.maddyhome.idea.vim.helper.subMode
 import com.maddyhome.idea.vim.listener.SelectionVimListenerSuppressor
+import org.jetbrains.annotations.ApiStatus
 import org.jetbrains.annotations.Contract
 import java.util.*
 import kotlin.math.ceil
@@ -83,6 +84,10 @@ object OptionsManager {
   val wrapscan = addOption(ToggleOption("wrapscan", "ws", true))
   val visualEnterDelay = addOption(NumberOption("visualdelay", "visualdelay", 100, 0, Int.MAX_VALUE))
   val idearefactormode = addOption(BoundStringOption(IdeaRefactorMode.name, IdeaRefactorMode.name, IdeaRefactorMode.select, IdeaRefactorMode.availableValues))
+  val ideastatusicon = addOption(BoundStringOption(IdeaStatusIcon.name, IdeaStatusIcon.name, IdeaStatusIcon.enabled, IdeaStatusIcon.allValues))
+
+  @ApiStatus.ScheduledForRemoval(inVersion = "0.58")
+  @Deprecated("please use ideastatusicon")
   val ideastatusbar = addOption(ToggleOption("ideastatusbar", "ideastatusbar", true))
 
   fun isSet(name: String): Boolean {
@@ -354,6 +359,11 @@ object OptionsManager {
     abbrevs += option.abbrev to option
     return option
   }
+
+  fun removeOption(name: String) {
+    options.remove(name)
+    abbrevs.values.find { it.name == name }?.let { option -> abbrevs.remove(option.abbrev) }
+  }
 }
 
 object KeyModelOptionData {
@@ -380,8 +390,11 @@ object SelectModeOptionData {
   const val key = "key"
   const val cmd = "cmd"
 
+  @ApiStatus.ScheduledForRemoval(inVersion = "0.57")
   @Deprecated("Please, use `idearefactormode` option")
   const val template = "template"
+
+  @ApiStatus.ScheduledForRemoval(inVersion = "0.57")
   @Deprecated("Please, use `ideaselection`")
   const val refactoring = "refactoring"
 
@@ -520,4 +533,13 @@ object LookupKeysData {
     // Vim: Insert next non-digit literally (same as <Ctrl-V>). Not yet supported (19.01.2020)
     "<C-Q>"
   )
+}
+
+object IdeaStatusIcon {
+  const val enabled = "enabled"
+  const val gray = "gray"
+  const val disabled = "disabled"
+
+  val name = "ideastatusicon"
+  val allValues = arrayOf(enabled, gray, disabled)
 }