+++ test/org/jetbrains/plugins/ideavim/ex/CommandParserTest.kt
@@ -18,16 +18,21 @@
 
 package org.jetbrains.plugins.ideavim.ex
 
+import com.maddyhome.idea.vim.VimPlugin
+import com.maddyhome.idea.vim.command.CommandState
+import com.maddyhome.idea.vim.ex.CommandParser
+import com.maddyhome.idea.vim.ex.CommandParser.EX_COMMAND_EP
+import com.maddyhome.idea.vim.ex.ExBeanClass
+import com.maddyhome.idea.vim.ex.ExCommand
 import com.maddyhome.idea.vim.ex.commands
+import com.maddyhome.idea.vim.ex.ranges.Ranges
 import junit.framework.TestCase
-import org.junit.Test
-import kotlin.test.assertEquals
+import org.jetbrains.plugins.ideavim.VimTestCase
 
 /**
  * @author Alex Plate
  */
-class CommandParserTest {
-  @Test
+class CommandParserTest : VimTestCase() {
   fun `test one letter without optional`() {
     val commands = commands("a")
     assertEquals(1, commands.size)
@@ -35,7 +40,6 @@ class CommandParserTest {
     assertEquals("", commands[0].optional)
   }
 
-  @Test
   fun `test without optional`() {
     val commands = commands("a_discovery")
     TestCase.assertEquals(1, commands.size)
@@ -43,11 +47,69 @@ class CommandParserTest {
     assertEquals("", commands[0].optional)
   }
 
-  @Test
   fun `test with optional`() {
     val commands = commands("a[discovery]")
     TestCase.assertEquals(1, commands.size)
     assertEquals("a", commands[0].required)
     assertEquals("discovery", commands[0].optional)
+  }
+
+  fun `test simple ex command execution`() {
+    val keys = commandToKeys(">>")
+    val before = "I ${c}found it in a legendary land"
+    val after = "        ${c}I found it in a legendary land"
+    doTest(keys, before, after, CommandState.Mode.COMMAND, CommandState.SubMode.NONE)
+  }
+
+  fun `test execute in disabled state`() {
+    val keys = commandToKeys(">>")
+    val before = "I ${c}found it in a legendary land"
+    val after = "I ${c}found it in a legendary land"
+    doTest(keys, before, after, CommandState.Mode.COMMAND, CommandState.SubMode.NONE) {
+      VimPlugin.setEnabled(false)
+    }
+  }
+
+  fun `test turn off and on`() {
+    val keys = commandToKeys(">>")
+    val before = "I ${c}found it in a legendary land"
+    val after = "        ${c}I found it in a legendary land"
+    doTest(keys, before, after, CommandState.Mode.COMMAND, CommandState.SubMode.NONE) {
+      VimPlugin.setEnabled(false)
+      VimPlugin.setEnabled(true)
+    }
+  }
+
+  fun `test turn off and on twice`() {
+    val keys = commandToKeys(">>")
+    val before = "I ${c}found it in a legendary land"
+    val after = "        ${c}I found it in a legendary land"
+    doTest(keys, before, after, CommandState.Mode.COMMAND, CommandState.SubMode.NONE) {
+      VimPlugin.setEnabled(false)
+      VimPlugin.setEnabled(true)
+      VimPlugin.setEnabled(true)
+    }
+  }
+
+  fun `test unregister extension`() {
+    val keys = commandToKeys(">>")
+    val before = "I ${c}found it in a legendary land"
+    val after = "        ${c}I found it in a legendary land"
+    var extension: ExBeanClass? = null
+    doTest(keys, before, after, CommandState.Mode.COMMAND, CommandState.SubMode.NONE) {
+      extension = EX_COMMAND_EP.extensions().findFirst().get()
+
+      // TODO: 08.02.2020 I'm sorry if your tests have been failed because of this code. Please update it properly
+      TestCase.assertNotNull(CommandParser.getInstance().getCommandHandler(ExCommand(Ranges(), "actionlist", "")))
+
+      @Suppress("DEPRECATION")
+      EX_COMMAND_EP.getPoint(null).unregisterExtension(extension!!)
+
+      TestCase.assertNull(CommandParser.getInstance().getCommandHandler(ExCommand(Ranges(), "actionlist", "")))
+    }
+    @Suppress("DEPRECATION")
+    EX_COMMAND_EP.getPoint(null).registerExtension(extension!!)
+
+    TestCase.assertNotNull(CommandParser.getInstance().getCommandHandler(ExCommand(Ranges(), "actionlist", "")))
   }
 }